+++ Runtime/Scripts/SelectionGroupDataLocation.cs
@@ -1,5 +1,6 @@
 ﻿namespace Unity.SelectionGroups
 {
+    //[TODO-sin: 2022-4-7] Remove this starting from 0.8.0-preview
     /// <summary>
     /// Indicates where the SelectionGroup data is stored
     /// </summary>