+++ SelectionGroups~/Packages/com.unity.selection-groups/CHANGELOG.md
@@ -4,8 +4,27 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.7.2-preview] - 2022-02-16
+## [0.7.4-preview] - 2022-04-15
 
+### Changed
+* deps: use com.unity.film-internal-utilities@0.14.2-preview 
+
+### Fixed
+* fix: support com.unity.quicksearch@3.x.x on Unity 2020.3.x
+
+## [0.7.3-preview] - 2022-04-11
+
+### Added
+* doc: update the group creation section in Getting Started and SelectionGroupsWindow pages
+
+### Changed
+* change: use a dropdown button to create a group or group from selection
+* deps: use com.unity.film-internal-utilities@0.14.1-preview 
+
+### Fixed
+* fix: the use of negative index relative to child count in GoQL
+
+## [0.7.2-preview] - 2022-02-16
 
 ### Added
 * api: make SelectionGroupToolAttribute public